refactor: update component import paths for improved organization
