+++ src/components/modules/workouts/ui/cards/DetailedWorkoutCard.tsx
