+++ src/components/modules/workouts/ui/cards/DetailedWorkoutExercise.tsx
