+++ src/components/modules/workouts/ui/cards/WorkoutCard.tsx
